start deleting files based on file properties in sql. Also ensure we delete original files when using external store

+++ src/Microsoft.Health.Dicom.Core/Features/Delete/DeleteService.cs
@@ -119,18 +119,18 @@ public class DeleteService : IDeleteService
                 {
                     try
                     {
-                        Task[] tasks = new[]
-                        {
-                            _fileStore.DeleteFileIfExistsAsync(deletedInstanceIdentifier.VersionedInstanceIdentifier.Version, deletedInstanceIdentifier.VersionedInstanceIdentifier.Partition.Name, cancellationToken),
+                        Task[] tasks = {
+                            _fileStore.DeleteFileIfExistsAsync(deletedInstanceIdentifier.VersionedInstanceIdentifier.Version, deletedInstanceIdentifier.VersionedInstanceIdentifier.Partition, deletedInstanceIdentifier.InstanceProperties.fileProperties, cancellationToken),
                             _metadataStore.DeleteInstanceMetadataIfExistsAsync(deletedInstanceIdentifier.VersionedInstanceIdentifier.Version, cancellationToken),
                             _metadataStore.DeleteInstanceFramesRangeAsync(deletedInstanceIdentifier.VersionedInstanceIdentifier.Version, cancellationToken),
                         };
 
-                        if (deletedInstanceIdentifier.InstanceProperties.OriginalVersion.HasValue)
+                        // only need to delete by "original watermark" to catch updates if not IDP
+                        if (!_isExternalStoreEnabled && deletedInstanceIdentifier.InstanceProperties.OriginalVersion.HasValue)
                         {
                             tasks = tasks.Concat(new[]
                             {
-                                _fileStore.DeleteFileIfExistsAsync(deletedInstanceIdentifier.InstanceProperties.OriginalVersion.Value,  deletedInstanceIdentifier.VersionedInstanceIdentifier.Partition.Name, cancellationToken),
+                                _fileStore.DeleteFileIfExistsAsync(deletedInstanceIdentifier.InstanceProperties.OriginalVersion.Value,  deletedInstanceIdentifier.VersionedInstanceIdentifier.Partition, deletedInstanceIdentifier.InstanceProperties.fileProperties, cancellationToken),
                                 _metadataStore.DeleteInstanceMetadataIfExistsAsync(deletedInstanceIdentifier.InstanceProperties.OriginalVersion.Value, cancellationToken),
                             }).ToArray();
                         }
